rtl8192ee: Fix lockup on start
Kernel commit 51d62f2f2c50 ("cfg80211: Save the regulatory domain with a lock")
caused a lock dependency to the cfg80211 version of this driver. It is fixed by
moving the regulatory domain initialization call so that it is outside the lock
in the probe routine.

See https://github.com/lwfinger/rtl8188eu/issues/386 for further discussion.
Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>